[ios][sdk-54][precompile] added support for USE_FRAMEWORKS with Expo modules (#39479)
# Why

Some Expo modules are including React Native Core header files to
override base RN classes in their public headers. This causes the swift
interop generation to fail, since the RN headers are not fully modular -
some include c++ code which fails when generating the module interface
for these modules.

This isn't a problem for the module itself - but for the modules or
targets that are using them. XCode will try to generate a public and
modular interface for the module using the Objective-c headers - but
when these contains c++ the process fails.

**NOTE:** Until we get a truly modular set of headers for React Native
using USE_FRAMEWORKS will fail with React Native precompiled binaries.
Building from source or building without USE_FRAMEWORKS does not fail
since header files are detected through search paths only - without any
checks for public/modularity.

This PR contains a workaround and not a final fix that we can implement
when the above is finished (it is in the works!).

Closes #39080

# How

This commit fixes this by disabling frameworks (forcing static linking)
of the targets that uses RN headers in their own public headers (Expo,
ExpoModulesCore, ReactAppDependencyProvider, expo-dev-menu).

To fix this I added an installer.rb class to expo-modules-autolinking
that overrides the Podfile and Installer and patches the necessary
modules. Also added `additionalFrameworkModulesToPatch` parameter to
use_expo_modules to be able to add additional framework modules to patch
when using USE_FRAMEWORKS. Not sure if this is needed, but it will be
very nice to have if someone needs it one day!

# Test Plan

Tested with new Expo app using SDK-54 and the changes from this commit:

✅ Build without precompiled RN, no USE_FRAMEWORKS
✅ Build without precompiled RN, USE_FRAMEWORKS=static 
✅ Build without precompiled RN, USE_FRAMEWORKS=dynamic 
✅ Build with precompiled RN, no USE_FRAMEWORKS
✅ Build with precompiled RN, USE_FRAMEWORKS=static
✅ Build with precompiled RN, USE_FRAMEWORKS=dynamic

# Checklist

- [x] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [x] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>
Co-authored-by: Kudo Chien <kudo@expo.dev>