[@expo/metro-config]: upgrade to `lightningcss@^1.30.1` and change to a caret range (#39506)
# Why

Upgrade `lightningcss` to the latest version (~3months old) and change
from using a `~` range to a `^` range.

`lightningcss` has shown itself to be a stable library, so there is no
need to limit the supported range to only patches. If anything, this has
caused some minor issues.

Due to `lightningcss` internal imports, the package does not operate
well as a nested dependency. This is the cause of the`Error: Cannot find
module '../lightningcss.<platform>'` error message
(https://github.com/expo/expo/issues/33340
https://github.com/expo/expo/issues/31775). It can become a nested
dependency when using two packages that both depend on different
versions of `lightningcss`. For example using both `expo` and
`@tailwindcss/postcss`

```
├─┬ @tailwindcss/postcss@4.1.13
│ └─┬ @tailwindcss/node@4.1.13
│   └── lightningcss@1.30.1
└─┬ expo@54.0.0-preview.16
  └─┬ @expo/metro-config@0.21.11
    └── lightningcss@1.27.0
```

The current "fix" is to use package resolutions/overrides, but this
wouldn't be needed if Expo supported the `^` range - has the closest
matching package would be hoisted to the top level.

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>