[web][expo][router] Check for isDomComponentContext instead of ReactNativeWebView.
Checking if window.ReactNativeWebview exists is not a valid method of
checking if the component is running in a DOM component. Some apps may
end up rendering an expo-web app inside of react-native-webview and need
the expo-web app to run exactly as it would in an ordinary browser.

We are running into this issue right now where we need to embed a third
party web app into our web app via an iframe. The third party web app
is an expo-web app that uses expo-router. Our mobile app uses
react-native-webview to render our web app. On Android, the iframe that
our web app uses to host the third party expo app has
window.ReactNativeWebView defined, and so the expo-web app in the iframe
completely breaks because it thinks it's running in a DOM Component when
we need it to behave as if it's running in any other web browser.

Co-Authored-By: Daniel Vicory <dvicory@gmail.com>