[ios] add build-props option to force pod target to be linked statically
After adding precompiled React Native binaries some npm packages that requires the app to be using frameworks (use_frameworks!) is exporting headers from React Native (which in the future will be fine) that causes build exceptions on iOS since the target ends up exporting a non-modular header (RN's headers arent 100% modular at the moment).

This is something that typically happens with the Firebase packages which require use_frameworks to be enabled. A solution is to link the offending pods directly (statically linking) with the app - leaving the other modules as frameworks.

This commit fixes the above problem by exposing a way to force some pod targets to be linked with the app. When linking statically the modular header requirement is no longer an issue.

To add f.ex the Firebase pods we can now use `expo-build-properties` and pass these libraries to the autolinking manager to ensure they are statically linked like in the below example:

```json
[
  "expo-build-properties",
   {
     "ios": {
     "useFrameworks": "static",
      "forceStaticLinking": ["RNFBApp", "RNFBAuth", "RNFBFirestore"]
     }
  }
]
```

This is now solved by adding the property `forceStaticLinking` which is a list of Pod target names for pods that should be patched to not build as frameworks.

This list will be saved to `Podfile.properties.json` and read back from within the `use_expo_modules!` function in `expo-modules-autolinking/scripts/ios/autolinking_manager.rb` and propagated to the installer override that patches the Pod target in `expo-modules-autolinking/scripts/ios/cocoapods/installer.rb`