[ios][background-task] fixed timing issue on startup
We've gotten a few reports from users of the background-task library getting errors like this:

```
Last Exception Backtrace:
0   CoreFoundation                  0x189cce8c8 __exceptionPreprocess + 164 (NSException.m:249)
1   libobjc.A.dylib                 0x186c457c4 objc_exception_throw + 88 (objc-exception.mm:356)
2   Foundation                      0x187bcce80 -[NSAssertionHandler handleFailureInMethod:object:file:lineNumber:description:] + 288 (NSException.m:252)
3   BackgroundTasks                 0x1b2aded80 -[BGTaskScheduler _handleSubmissionWithoutRegistrationForTaskRequest:error:] + 204 (BGTaskScheduler.m:289)
4   BackgroundTasks                 0x1b2adf130 -[BGTaskScheduler _unsafe_submitTaskRequest:error:] + 664 (BGTaskScheduler.m:301)
5   BackgroundTasks                 0x1b2adec70 -[BGTaskScheduler submitTaskRequest:error:] + 168 (BGTaskScheduler.m:267)
6   StubHub                         0x102deba44 specialized static BackgroundTaskScheduler.tryScheduleWorker() + 472 (BackgroundTaskScheduler.swift:67)
```

This is caused by a timing issue where `expo-task-manager` calls `taskDidRegister` in `BackgroundTaskConsumer` which in turn ends up trying to register a BGTask - and in some circumstances (different threads) the BGTaskScheduler isn't done registering the handler yet.

This fixes this by adding a registration gate class that lets us wait until the BGTaskScheduler is done registering its handler.

This should fix the timing issue:

Closing #39406