ResultArray length was encoded on 1 byte instead of 4
I had requests returning corrupted json data when their result length was over 255 chars. The result contained my json into the memory, but the length was wrong. After digging, I saw that the length (which is supposed to be encoded into the 4 first bytes) was encoded in a short (1 first byte). My PR fixes this problem