[core][ios] Fix NSURL to JS conversion returnig nil (#39567)
# Why

I have noticed that for expo-video events the `uri` field of the event
payload is nil on the JS side, even though the value is passed into the
`emit` method of the SharedObject.

It looks like `EXJSIConversions` for objects do not support URL/NSURL
values.

# How

Added a conversion for URL and NSURL that returns the `absoluteString`
value of the url.

# Test Plan

Tested in BareExpo on iOS

---------

(cherry picked from commit b0bac89e439d212eea45ce4840c540c6201121c4)