chore: Update issue templates with larger note (#39798)
This is just a proposal, which I believe I've mentioned to @alanjhughes
in DMs. I've seen similar issues on the
[`gql.tada`](https://github.com/0no-co/gql.tada) and `urql` repos. To
experiment, a few months back we added a markdown "important" callout to
the issue template, which is _very_ visible when opening the issue
creation modal.

Adding this notice and getting people's attention in this way improved
the issue quality there dramatically. This is anecdata, but if we think
this is safe to try out, it's relatively low risk to revert or tweak as
needed.

Copy here isn't final. I've made a tweak to clarify that _reducing_ an
app to a minimal reproduction is a valid way to provide a reproduction.
I've also updated the CLI issue template to have an input field rather
than text are for the repro, since that seems to have been a typo

---------

Co-authored-by: Hassan Khan <hassan@expo.dev>