refactor(server): Prune `@expo/server` public API to not contain vendor export and refactor (#39894)
# Why

We'd like to clear out the `@expo/server` public API to be user-usable
and public, preparing it to move to the `expo-server` name and becoming
a public API. As part of this, not only do types have to move around
appropriately, we'll also need to hide away `createRequestHandler`.

Main motivation:
- `@expo/server` should be importable in server-side code
side-effect-free
- Main entrypoint should not lead to utility/vendor code
- abstract `createRequestHandler` should be placed elsewhere
(`@expo/server/vendor/abstract`)

# How

- Remove `ExpoError` from public export
- Move `createRequestHandler` to `vendor/abstract`
- Extract common node and workerd env code into abstract factory
functions
- Refactor middleware types to be exposed (so e.g. `unstable_settings`
or `ImmutableRequest` can be typed)
- Hide away `MiddlewareModule` and make it part of abstract vendor type
- **Behaviour change:** Remove intermediate function in `vendor/http`
for manifest loading
- Vendor manifest types into `src/manifest.ts` (these won't be
user-exposed and instead be re-exposed from `/private`)
- Expose all of `src/types.ts` (after prune) as public API types
- Move assertion call for node env to `vendor/environment/node` and mark
as `sideEffects: false`

> [!NOTE]
> I'm currently assuming that while this is technically a breaking
change, we can safely backport this to SDK 54. The reason for this is
that `createRequestHandler` isn't documented, and that we may bump the
major version of this package. It's not a classic public API yet, so we
may even rename it to `expo-server`. Hence, the merge commit resulting
from this PR is backportable.

# Test Plan

- Existing CI should check API-compatibility

# Checklist

- [x] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)