[core] Fix SharedObject getting destroyed after a code change in `useReleasingSharedObject` deps (#39753)
# Why

Fixes https://github.com/expo/expo/issues/39427

The (evil 👹) `useReleasingSharedObject` hook will release a shared
object after a fast refresh caused by a change in its own dependencies.
Causing this error
```
Unable to find the native shared object associated with given JavaScript object
```

Here is the code that's relevant in the fix

The crucial part is that this code is called when a fast refresh happens
and when the dependencies of the hook change.

```
    if (!newObject || !dependenciesAreEqual) {
      objectRef.current?.release();
      newObject = factory();
      objectRef.current = newObject;
      previousDependencies.current = dependencies;
    } else {
      isFastRefresh.current = true;
    }
```

Consider the following example that currently works - we change
something in our project, which causes a fast refresh.

We enter the `else` block since the deps are the same, the refresh is
marked as fast refresh. Then we enter the `return` of the `useEffect`
locaded below and we know that the current refresh is fast and that we
don't have to release the object.

Now consider the following:
We change the dependencies of the object by editing the code (i.e. we
fix a typo in the video source string). We enter the `if` block since
the dependencies are different. We correctly release the current object
and create a new one. The change also causes a fast refresh, since the
code of the app has been changed. We don't mark the change as a fast
refresh because currently we always assume that the changes come from a
state change and not code change. The fast refresh still happens and the
return block of `useEffect()` is called destroying the newly created
object and causing errors.


# How

We need to mark the problematic change as fast refresh. We can't just
put `isFastRefresh.current = true;` outside the `else` block, because it
is also called when dependencies change due to state change which
happens without a fast refresh.

Therefore I have added another `useMemo` hook that doesn't have
dependencies. This one will be recalculated on fast refreshes (and
initial load but that one doesn't matter, since it is cancelled out by
the useEffect call).

# Test Plan

Tested in BareExpo on Android and iOS