[video][android] Fix service premission being always present in the manifest (#40074)
# Why

Fixes https://github.com/expo/expo/issues/39988

The foreground service permission should've been removed from the
expo-video manifest in https://github.com/expo/expo/pull/38980, which
moved that functionality into the config plugin, but it wasn't, which
means that people, who don't need that permission still have in the app
manifest.

# How

Remove the permission from the manifest, the config plugin already
handles it.

# Test Plan

Tested in a new project to see if the permission is getting added when
necessary.