[expo-audio] Align pitch correction on iOS/Android
Currently Expo-audio does pitch correction (when playback rate is not 1.0) out of the box on iOS even though the shouldCorrectPitch property of the Player is false by default. On Android this works "as expectced", when shouldCorrectPitch is not set or false, it will not correct the pitch.

This commit fixes this by changing the defaults a bit:

- The `shouldCorrectPitch` property is now default to true to align iOS/Android
- The quality level on iOS is changed so that the one NOT doing pitch correction will be the default when shouldCorrectPitch is false