fix(cli): Prevent externals resolution logic from being applied to unreachable Node modules (#40247)
Resolves #40192

# Why

This is more of a temporary workaround rather than a solution. It's
fundamentally unsafe for us to apply externals to "random" Node modules.
They may not be hoisted nor reachable by the bundle and where it's being
executed. The only safe way to achieve this in development is to detect
modules that are safe to run in Node.js and then to externalise them
either if:
- their origin module (the requestor) is not a Node module itself
- if it's reachable via Node resolution in the project root

Externalising any module can cause modules that aren't hoisted (or
reachable due to isolated dependencies) to cause errors during
development-time, since Node resolution in Node.js won't be able to
resolve them. We should align the externals logic to the "autolinking
resolution" resolver so we can filter it to specific modules and/or
delete this entirely in the future.

# How

For now, we can resolve the externalised modules properly by checking if
they're resolvable from the project root.

Generally, the externals logic looks to me like a dev-time-only
optimisation. We should be safe to skip it if we can detect that the
module we're trying to request isn't "reachable". This can be done by
overriding the `originModulePath` with a path based on the project root,
which is what we're already doing in a few other cases.

# Test Plan

- **TODO**: cc @Ubax for if we have a reliable repro

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
-->

- [x] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)