[expo/core][android] Fix `androidNavigationBar.enforceContrast` app config property not working (#40263)
# Why

> I was investigating the react-conf app, and discovered that our
enforceContrast property, which is based only on editing styles.xml is
not working.
I’ve traced back the issue to these lines in react-native

[https://github.com/facebook/react-native/blame/50273758510a6d756494d05dc91055516c2[…]droid/src/main/java/com/facebook/react/views/view/WindowUtil.kt](https://github.com/facebook/react-native/blame/50273758510a6d756494d05dc91055516c2a6aca/packages/react-native/ReactAndroid/src/main/java/com/facebook/react/views/view/WindowUtil.kt#L111-L114)
Setting isNavigationBarContrastEnforced = true overrides all settings
from the theme.

# How

Moved all edge-to-edge logic into the `EdgeToEdge` package in
`expo-modules-core`, I felt like it's not worth cluttering the
ReactActivityDelegateWrapper with it.
After we apply the edge to edge method, we call
`activity?.enforceNavigationBarContrastFromTheme()` to restore
`isNavigationBarContrastEnforced` to the value set in styles (this value
comes from our config plugin)

# Test Plan

Tested in Android 11.0 and 16.0 in both bare-expo and a fresh test app.