[doctor] Improve autolinking check output for isolated and corrupted installations (#40279)
# Why

The messages in this check basically have two problems right now:
- It's not obvious to users that we support isolated dependencies and
that the check is aware of them
- Corrupted installations aren't always detected and the action the
message for them advises users to take is incomplete

The first problem basically is that we may present mixed store and
non-store paths. This could be misunderstood by users that are trying
out isolated dependencies that this check isn't aware of isolated
dependencies, which isn't correct.

The second problem is specifically that we didn't always detect
corrupted installations, but that we also didn't give actionable advice
for a bug in Bun with isolated dependencies. We have to tell users
explicitly to remove the lockfile if the first advice doesn't work.

# How

- If we're list any path that is a store path, we should display the
realpath i.e. what the origin path of the dependency is symlinked to
- If `version` isn't provided/scanned by autolinking, we must load it
manually when checking for corrupted installations
- When telling users how to fix corrupted installations, if the first
step fails (deleting `node_modules`) we should tell them to also delete
the lockfile
- I've chosen to hide some symlinked path, since the output can
otherwise get really noisy. It's always accurate for us to just say "a
different installation" since we know the `path`s differ

This is what all of these changes look like together in practice:

<img width="1332" height="603" alt="Screenshot 2025-10-09 at 03 44 47"
src="https://github.com/user-attachments/assets/dea84945-7e2e-4741-9dd9-b130f7aff2a4"
/>

# Test Plan

You can reproduce a corrupted installation pretty quickly:
- create a project with `bunx create-expo`
- cd into it and run `bun install --linker=isolated`

The old messages will not show that this is a corrupted installation,
but the doctor output will speak for itself. When this PR is applied the
improved messages (as seen in the above screenshot) will be displayed
instead.

When Bun is in this state, only deleting the lockfile and `node_modules`
will restore the correct installation behaviour. (No matter if that's
followed by `--linker=isolated` or `--linker=hoisted`). This Bun bug is
specific to an initial isolated on top of a hoisted installation
specifically.

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
-->

- [x] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)