[android][blur][1/n] BlurView audit - introduce new Blur API (#39990)
# Why

Currently, background blur with `expo-blur` is marked as experimental
mostly due to bad blur performance (we used to have some flickering
issues, which should now be resolved).

The underlying `DimezisBlurView` library recently got an update, which
greatly improves performance on Android 12.0+, but requires users to
wrap the blurred component in a `BlurTarget`

# How

This PR updates the `DimezisBlurView` library to the newest version, and
introduces the new API to the JS size. Now to blur on Android you have
to do something like this:

```
<BlurTargetView>
<Image source="backgroundImage.jpg">
<BlurTargetView>
<BlurView>
<Text>Content behind me is blurred!</Text>
</BlurView>
```

For users, who don't upgrade to the new method, using the current method
will lead to a result equivalent to having set `experimentalBlurMethod`
to `none`.

For Android versions below 12 the old, less efficient blur method is
still used. This issue is addressed in one of the next PRs.

# Test Plan

Tested in BareExpo on Android 16, 9.0 and iOS