[expo-audio] Removed unneeded call to setIsAudioActiveAsync(true) (#40299)
# Why

Previously we added a call to setIsAudioActiveAsync(true) in the example
screen before enabling the lock-screen otherwise it wouldn't become
visible.

# How

This commit removes the call to setIsAudioActiveAsync before playing
sound in the AudioControlsScreen. Also added a catch clause to get help
if something is wrong.

# Test Plan

Tested in Bare-Expo + client project:
✅ Android
✅ iOS

# Checklist

- [x] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).