[expo-audio][ios] fixed issue with configured buttons not showing up correctly (#40275)
# Why

On the lock screen on iOS the lock screen buttons wasn't showing up
correctly the first time you enabled it in an audio session, but next
time they started working.

This commit fixes this.

# How

The issue was that we tried initializing the remote command center
before the app was made the active now-playing app since the
setupRemoteCommands() method was called in the ctor/init of the
MediaController and not when setting the player as the active player.

Fixes:
- Added initialization of the remote command center after we've
activated and set up the player.
- Removed the init method from the MediaController class
- Been a bit more explicit about cleaning up targets and disabling
unwanted commands.
- Added a fix to the category options that supports Xcode 26

# Test Plan

✅ BareExpo

# Checklist

- [x] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).