[video][ios] Fix crashes when loading HLS video tracks (#40265)
# Why

When rapidly replacing video sources and releasing the players,
sometimes the tracks loading task will outlive the VideoPlayerItem
and/or the VideoPlayer, causing crashes upon conversion when sending an
event. This happens only for HLS sources - we use a custom track loading
solution for HLS.

Here is a loop that would cause a crash within ~5-10 seconds.

```
setInterval(() => {
  playerRef.current.release();
  playerRef.current = createVideoPlayer(null);
  playerRef.current.replaceAsync(hlsSource);
}, 500);
```

# How

Cancel the loading task on deinit and when a new source begins loading.

# Test Plan

Tested in bare expo on iOS simulator. Ran the loop above for about 10
minutes and didn't experience any crashes.