[fingerprint] prevent package.json fingerprint from ExpoConfigLoader (#40320)
# Why

fixes #39683

# How

if people load project's package.json in app.config.js, the package.json
will appear in the fingerprint from ExpoConfigLoader. that's not the
intention for ExpoConfigLoader. for package.json, we only include the
"scripts" section by default rather than the whole file.

this pr tries to exclude the package.json from the ExpoConfigLoader
phase.