chore(expo): Remove all react canary support now that React 19 is the default (#40386)
# Why

- We no longer need canary builds of React to develop React Server
Components or owner stack. We may need to add this back in the future
when some new React feature comes out, but for now it's just adding a
layer of untested complexity that no one is actively maintaining and
currently adds nothing.
- The flag will now warn that nothing is happening and that you can
remove the flag.

# Test Plan

- Ran the three RSC tests on iOS, Android, and web. All appeared fine.
We were close last round but the sibling issue in React suspense was
breaking native RSC.

---------

Co-authored-by: Phil Pluckthun <phil@kitten.sh>