fix(expo-modules-core): fix RSC and add tests (#40388)
# Why

- The RSC code paths are special because they run native code in server
mode (e.g. `typeof window === 'undefined'`). When we add new code to
core, it's easy to create subtle issues where we create a module that's
only available in native clients but not native servers.
- These issues are subtle but can add up, and only get caught in the
tests of other packages such as expo-linking. So this PR adds a basic
and fast test to check if any warnings or errors are logged when
importing expo modules core in a native server environment.
- fix regression from https://github.com/expo/expo/pull/39726

# Test Plan

- Add `yarn test:rsc` which should run in check-packages to ensure any
changes to core will not break RSC for native.

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>