[android][camera] Allow opt out of barcode scanning deps (#40445)
# Why
MLKit adds significant size to the apk which is not necessary when you are not using it. 

# How
Switch the barcode scanning dependencies to `compile-only` which will cause them to not be included in the apk. In debug builds there is about a 10mb size difference when MLKit is excluded in the sandbox app. Couldn't check release because I ran into some expo-router issues. I wouldn't expect as large a difference but there will be a decrease 

# Test Plan
Sandbox