[autolinking] fix globMatchFunctorAllAsync to check if globbed path is a file to prevent errors (#40437)
# Why

<!--
Please describe the motivation for this PR, and link to relevant GitHub
issues, forums posts, or feature requests.
-->

The `globMatchFunctorAllAsync` utility reads files (paths) that match a
certain functor, but does not check if the path is actually a file. This
means that if a globbed path is a directory matching the `globPattern`
and satisfying the `matchFunctor`, it will result in the `EISDIR` error
being thrown when trying to `fs.readFile` such a path.

An example when this fails is, when a package in `node_modules` contains
a directory named ~ `<anything>.{t,j}sx?` and the expo autolinking
command is run.

# How

<!--
How did you build this feature or fix this bug and why?
-->

If the path is not a file (`fs.stat(...).isFile()`), then `continue` the
loop.

# Test Plan

<!--
Please describe how you tested this change and how a reviewer could
reproduce your test, especially if this PR does not include automated
tests! If possible, please also provide terminal output and/or
screenshots demonstrating your test/reproduction.
-->

Tested in a locally created project that contains a directory:
`node_modules/<test package>/**/<some directory>.js`.

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
-->

- [x] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)