[android][expo-audio] removed use of setShowActionsInCompactView when building custom layout
When using custom layouts to build the media lock screen notifcation, we don't need to call the `setShowActionsInCompactView` when building the notification, since it will use the order in the custom layout anyway.

This commit fixes this by removing the call to setShowActionsInCompactView. This was previously causing errors in Sentry like the following:

```
Couldn't inflate contentViewsjava.lang.IllegalArgumentException: setShowActionsInCompactView: action 0 out of bounds (max -1)
```