[ios][pods] fixed issue with unstable ids.
# Why

We previously added a fix to make sure React codegen is ran correctly while waiting for a fix in React Native to be released.

This temporary fix can be found here:
https://github.com/expo/expo/pull/40219

In a very rare situation this script caused some UUIDs to clash in the produced Pods.pbxproj file.

This could be reproduced by:

```bash
npx create-expo-app@latest --template blank
....
npx expo add expo-audio
```

Try to run and observe it fails. Open the Workspace in Xcode and observe that it is unable to open the Pods project (you can also try opening the pods project directly - it will fail).

Xcode fingerprints a PBX object from stable inputs that include the build-phase index. When you delete/insert to move the phase before “Compile Sources,” you change that index; after adding a new pod the whole object graph shifts, and your phase’s hash can (deterministically) collide with the Pods project’s rootObject UUID.

# How

The problem was that the script we injected was injected at the wrong time in the Cocoapods installer timeline.

This commit fixes this by moving it from `run_podfile_post_install_hooks` -> `perform_post_install_actions`.

# Testing

Do the same as the reproduction, now expect the Pods project to load correctly.