[ui] first attempt to add base view (#40492)
# Why

we have much duplicated code when adding a new view

# How

this is the effort for swift part. trying to add a base view and base
view props that most views can reuse

- typescript code sharing is not covered in this pr
- views with AsyncFunctions are not supported yet. need to figure out a
way to get correct ViewType for `ExpoSwiftUI.ViewDefinitionBuilder`. or
maybe we should add our own ViewDefinitionBuilder in expo-ui

---------

Co-authored-by: Tomasz Sapeta <tsapeta@expo.dev>