[video] Add `seekTolerance` and `scrubbingModeOptions` properties (#40203)
# Why

Media3 ExoPlayer 1.8.0 has introduced scrubbing mode, which improves
scrubbing performance on Android. We can also introduce a custom
scrubbing mode on iOS, which reduces the number of `AVPlayer.seek`
calls.

It is also helpful to have a `seekTolerance` property, which allows for
further fine-tuning of scrubbing performance.

Suggested in https://github.com/expo/expo/issues/39831

# How

Add `seek tolerance` and `scrubbingModeOptions` properties, updated the
docs, included a new video, optimized for seeking for the demo (the
source consists of only keyframes making seeking very fast. Such source
wouldn't be used in a playback scenario, but would be a good choice for
a situation where the video playback position is supposed to be in sync
with some UI element, for example a video that progresses as the user
scrolls the page)

# Test Plan

Tested in BareExpo on Android and iOS 


https://github.com/user-attachments/assets/a56373db-efef-4cba-a77b-8b4504ecd4ec