fix(cli,metro-config,expo): Fix `asyncRequireModulePath` absolute path causing transform cache pollution (#40584)
# Why

The `asyncRequireModulePath` has been set to a resolved path for a
while. If it's set to a module request — as intended by Metro for
`asyncRequireModulePath` and `assetRegistryPath` — it pollutes the
transform cache. Specifically, this path is inserted in the transform
worker (typically, during the `collect-dependencies` step for example)
to replace async imports.

However, Metro doesn't have a built-in mechanism to prevent cache
pollution. It's expected that this path is inserted as-is.

When we made Expo ready for isolated installations, we didn't have to
adjust the `assetRegistryPath`'s resolution path, since it was already
set to a module request that's aliased by our Metro config resolution
(via a `/^@react-native\/assets-registry\/registry(\.js)?$/` check)

However, we didn't consider that this is also necessary for the
`asyncRequireModulePath`. Specifically,
- we can't set `asyncRequireModulePath` to any resolved/absolute path,
and it **shouldn't change** so cached transformer outputs are stable
- we should add a resolution rule for `asyncRequireModulePath` so it can
resolve even with isolated installations or when the module request set
on `asyncRequireModulePath` isn't reachable from a module (it must be
resolvable from any module)

We broke these requirements previously and pulled
`asyncRequireModulePath` out of the cache key, which caused cache
mismatches to be possible when `asyncRequireModulePath`'s resolved path
wasn't present anymore in the module system.

# How

- Move `async-require` module from `@expo/metro-config` to
`expo/internal/async-require-module`
- **NOTE:** This should be temporary until we can move the resolvers to
`@expo/metro-config`. For now, when we can't find `expo`, we default to
Metro's default
- Remove `@expo/metro-config/src/async-require.ts`
- **NOTE:** This is safe because we're altering the Metro cache key in
this PR
- Add `expo/internal/async-require-module` pointing to
`expo/src/async-require/asyncRequireModule.ts`
- Add manual `resolve-from` resolution from
`config.transformer.asyncRequireModulePath` request to its resolved path
(when it can resolve)
- Fix `getExpoMetroRuntimeOptional` resolution fallback to work when
`expo` isn't installed
- Rename `requestStableAssetRegistry` to `requestStableConfigModules`

> [!NOTE]
> We should aim to move the `async-require-module` back to
`@expo/metro-config`. However, that's only really "clean" if we move the
resolver pipeline and our Metro modifications there too

# Test Plan

- Updated resolver tests to check that asset registry and async require
resolution work as expected
- Manually tested by picking changes to `sdk-54` and testing it against
a new project that contains an async `import()` and observing the
resolution in `requestStableConfigModules`

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
-->

- [x] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)